pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-slop-detector"
version = "0.1.0"
description = "Binary text classifier for AI-vs-human text detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "evaluate>=0.4.1",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
onnx = [
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "ai-slop-detector[onnx]",
]

[tool.setuptools.packages.find]
include = ["detector*"]

[tool.ruff]
target-version = "py310"
line-length = 120
select = ["E", "F", "I", "W", "UP"]

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
]
